exception handling springboot

+++ java-code/datajpa/src/main/java/com/neueda/datajpa/service/ItemServiceImpl.java
@@ -1,6 +1,8 @@
 package com.neueda.datajpa.service;
 
 import com.neueda.datajpa.entity.Item;
+import com.neueda.datajpa.exception.InvalidParameterException;
+import com.neueda.datajpa.exception.ItemNotFoundException;
 import com.neueda.datajpa.repository.ItemRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -25,23 +27,28 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public Item getItemById(Long id) {
-        return repository.findById(id).orElse(null);
+    public Item getItemById(Long id) throws ItemNotFoundException {
+        return repository.findById(id)
+                .orElseThrow(()->new ItemNotFoundException("Item with Id: "+id+" not found"));
     }
 
     @Override
-    public Item updateItem(Long id, Item update) {
+    public Item updateItem(Long id, Item update) throws ItemNotFoundException {
         return repository.findById(id)
                 .map(existingItem-> {
                     existingItem.setName(update.getName());
                     existingItem.setPrice(update.getPrice());
                     return repository.save(existingItem);
-                }).orElse(null);
+                })
+                .orElseThrow(()->new ItemNotFoundException("Item with Id: "+id+" not found"));
 
     }
 
     @Override
-    public void deleteItem(Long id) {
+    public void deleteItem(Long id) throws ItemNotFoundException {
+        if(!repository.existsById(id)){
+            throw new ItemNotFoundException("Item with Id: "+id+" not found");
+        }
         repository.deleteById(id);
     }
 }
